mend

+++ src/PSDataIO.jl
@@ -42,8 +42,8 @@ mutable struct PSTopology
     wind_loc::Vector{Int64} # bus ids of wind farms 
     bus_areas::Vector{Int64} # area index for each bus
     bus_type::Vector{Int64} # type of bus 1: PQ, 2: PV, 3: Slack
-    bus_Gs::Vector{Int64} # shunt conductance
-    bus_Bs::Vector{Int64} # shunt susceptance
+    bus_Gs::Vector{Float64} # shunt conductance
+    bus_Bs::Vector{Float64} # shunt susceptance
     gen_pmax::Vector{Float64} # generator production limit
     gen_pmin::Vector{Float64} # generator minimum production
     gen_qmax::Vector{Float64} # generator reactive power max
